requirements.txt
fastapi
uvicorn
google-genai
duffel-api
python-dotenv
pydantic
jinja2
